backend/requirements.txt
psutil
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
GPUtil
pyinstaller